src/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
drf_yasg==1.21.4
jsonschema==4.17.3
psycopg2==2.9.5
requests==2.28.1
rest_framework_simplejwt==0.0.2
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
python_dateutil==2.8.2
scikit_learn==1.2.2